Turned camera into a component and cleaned up clutter
